requirements.txt
fastapi
llama_cpp_python
ruff
uvicorn[standard]
pydantic
httpx
